still working on adding clean architecture to the app
