+++ app/src/main/java/ru/startandroid/develop/notebook/core/Constants.kt
@@ -1,0 +1,6 @@
+package ru.startandroid.develop.notebook.core
+
+/* Файл содержит все константы приложения */
+object Constants {
+
+}
