+++ app/src/main/java/ru/startandroid/develop/notebook/core/extensions/ContextExt.kt
@@ -1,0 +1,8 @@
+package ru.startandroid.develop.notebook.core
+
+import android.content.Context
+import android.widget.Toast
+
+fun Context.toast(message: String, isLong: Boolean = false) {
+    Toast.makeText(this, message, if (isLong) Toast.LENGTH_LONG else Toast.LENGTH_SHORT).show()
+}
