+++ app/src/main/java/ru/startandroid/develop/notebook/data/converters/NoteEntityToDomainConverter.kt
@@ -1,0 +1,12 @@
+package ru.startandroid.develop.notebook.data.converters
+
+import ru.startandroid.develop.notebook.data.enitities.NoteEntity
+import ru.startandroid.develop.notebook.domain.model.NoteDomainModel
+
+fun NoteEntity.toDomain() = NoteDomainModel(
+    id = id,
+    header = header,
+    description = description,
+    timeStamp = timeStamp,
+    createdDateFormatted = createdDateFormatted
+)
