+++ app/src/main/java/ru/startandroid/develop/notebook/data/db/NoteDao.kt
@@ -1,0 +1,25 @@
+package ru.startandroid.develop.notebook.data.db
+
+import androidx.room.*
+import kotlinx.coroutines.flow.Flow
+import ru.startandroid.develop.notebook.data.enitities.NOTE_TABLE_NAME
+import ru.startandroid.develop.notebook.data.enitities.NoteEntity
+
+@Dao
+interface NoteDao {
+
+    @Query("SELECT * FROM $NOTE_TABLE_NAME")
+    fun getNotes() : Flow<List<NoteEntity>>
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertNote(note: NoteEntity)
+
+    @Update
+    suspend fun update(note: NoteEntity)
+
+    @Delete
+    suspend fun delete(note: NoteEntity)
+
+    @Query("DELETE FROM $NOTE_TABLE_NAME")
+    suspend fun deleteAllNotes()
+}
