+++ app/src/main/java/ru/startandroid/develop/notebook/data/sharedpreferences/PreferenceHelper.kt
@@ -1,0 +1,8 @@
+package ru.startandroid.develop.notebook.data.sharedpreferences
+
+interface PreferenceHelper {
+
+    fun saveString(key: String, value: String)
+
+    fun getString(key: String): String?
+}
