+++ app/src/main/java/ru/startandroid/develop/notebook/data/sharedpreferences/SharedPreferencesKeys.kt
@@ -1,0 +1,6 @@
+package ru.startandroid.develop.notebook.data.sharedpreferences
+
+object SharedPreferencesKeys {
+
+    const val USER_SELECTED_THEME_MODE_KEY = "USER_SELECTED_THEME_MODE_KEY"
+}
