+++ app/src/main/java/ru/startandroid/develop/notebook/di/PreferencesModule.kt
@@ -4,13 +4,13 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import ru.startandroid.develop.notebook.sharedpreferences.PreferenceHelper
-import ru.startandroid.develop.notebook.sharedpreferences.PreferenceHelperImpl
+import ru.startandroid.develop.notebook.data.sharedpreferences.PreferenceHelper
+import ru.startandroid.develop.notebook.data.sharedpreferences.PreferenceHelperImpl
 
 @Module
 @InstallIn(SingletonComponent::class)
 interface PreferencesModule {
 
     @Binds
-    abstract fun providePreferenceHelper(preferenceHelperImpl: PreferenceHelperImpl): PreferenceHelper
+    fun providePreferenceHelper(preferenceHelperImpl: PreferenceHelperImpl): PreferenceHelper
 }
