+++ app/src/main/java/ru/startandroid/develop/notebook/domain/model/NoteDomainModel.kt
@@ -1,0 +1,9 @@
+package ru.startandroid.develop.notebook.domain.model
+
+data class NoteDomainModel(
+    val id: Int,
+    val header: String,
+    val description: String,
+    val timeStamp: Long,
+    val createdDateFormatted: String,
+)
