--- app/src/main/java/ru/startandroid/develop/notebook/model/NoteDao.kt
@@ -1,23 +1,0 @@
-package ru.startandroid.develop.notebook.model
-
-import androidx.room.*
-import kotlinx.coroutines.flow.Flow
-
-@Dao
-interface NoteDao {
-
-    @Query("SELECT * FROM $NOTE_TABLE_NAME")
-    fun getNotes() : Flow<List<NoteEntity>>
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertNote(note: NoteEntity)
-
-    @Update
-    suspend fun update(note: NoteEntity)
-
-    @Delete
-    suspend fun delete(note: NoteEntity)
-
-    @Query("DELETE FROM $NOTE_TABLE_NAME")
-    suspend fun deleteAllNotes()
-}
