--- app/src/main/java/ru/startandroid/develop/notebook/model/NoteDatabase.kt
@@ -1,10 +1,0 @@
-package ru.startandroid.develop.notebook.model
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-
-@Database(entities = [NoteEntity::class], version = 6,exportSchema = false)
-abstract class NoteDatabase : RoomDatabase() {
-
-    abstract fun noteDao() : NoteDao
-}
