+++ app/src/main/java/ru/startandroid/develop/notebook/screens/main/ui/adapter/NoteAdapter.kt
@@ -1,0 +1,21 @@
+package ru.startandroid.develop.notebook.screens.main.ui.adapter
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.ListAdapter
+import ru.startandroid.develop.notebook.data.enitities.NoteEntity
+import ru.startandroid.develop.notebook.databinding.NoteItemBinding
+
+class NoteAdapter(
+    private val listener: NoteItemClickListener
+) : ListAdapter<NoteEntity, NoteViewHolder>(NoteDiffCallback()) {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): NoteViewHolder {
+        val binding = NoteItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        return NoteViewHolder(binding, listener)
+    }
+
+    override fun onBindViewHolder(holder: NoteViewHolder, position: Int) {
+        holder.bind(getItem(position))
+    }
+}
