--- app/src/main/java/ru/startandroid/develop/notebook/sharedpreferences/PreferenceHelper.kt
@@ -1,8 +1,0 @@
-package ru.startandroid.develop.notebook.sharedpreferences
-
-interface PreferenceHelper {
-
-    fun saveString(key: String, value: String)
-
-    fun getString(key: String): String?
-}
