--- app/src/main/java/ru/startandroid/develop/notebook/sharedpreferences/SharedPreferencesKeys.kt
@@ -1,6 +1,0 @@
-package ru.startandroid.develop.notebook.sharedpreferences
-
-object SharedPreferencesKeys {
-
-    const val USER_SELECTED_THEME_MODE_KEY = "USER_SELECTED_THEME_MODE_KEY"
-}
