--- app/src/main/java/ru/startandroid/develop/notebook/utils/AppThemeModes.kt
@@ -1,11 +1,0 @@
-package ru.startandroid.develop.notebook.utils
-
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-@kotlinx.parcelize.Parcelize
-enum class AppThemeModes : Parcelable {
-    DARK_MODE,
-    LIGHT_MODE,
-    SYSTEM_ADAPT
-}
