--- app/src/main/java/ru/startandroid/develop/notebook/utils/Constants.kt
@@ -1,6 +1,0 @@
-package ru.startandroid.develop.notebook.utils
-
-/* Файл содержит все константы приложения */
-object Constants {
-
-}
