--- app/src/main/java/ru/startandroid/develop/notebook/view/noteadapter/NoteAdapter.kt
@@ -1,24 +1,0 @@
-package ru.startandroid.develop.notebook.view.noteadapter
-
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import ru.startandroid.develop.notebook.databinding.NoteItemBinding
-import ru.startandroid.develop.notebook.model.NoteEntity
-
-class NoteAdapter(
-    private val listener: NoteItemClickListener
-) : ListAdapter<NoteEntity, NoteViewHolder>(NoteDiffCallback()) {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): NoteViewHolder {
-        val binding = NoteItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return NoteViewHolder(binding, listener)
-    }
-
-    override fun onBindViewHolder(holder: NoteViewHolder, position: Int) {
-        holder.bind(getItem(position))
-    }
-}
