+++ app/src/main/res/layout/note_add_edit_fragment.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context=".view.NoteAddEditFragment"
+    tools:context=".screens.addedit.ui.view.NoteAddEditFragment"
     android:background="@color/note_add_edit_fragment_background_color">
 
     <androidx.appcompat.widget.LinearLayoutCompat
