+++ app/src/main/res/layout/note_main_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.NoteMainFragment">
+    tools:context=".screens.main.ui.view.NoteMainFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/note_main_recyclerView"
